requirements.txt
absl-py
gin-config
numpy
torch
pytorch-lightning
pytorch-metric-learning
umap-learn
tensorboard
matplotlib
pandas
plotly
